restructured the heat_generator classes to multiple files
