+++ src/districtheatsim/gui/MixDesignTab/calculate_mix_thread.py
@@ -1,7 +1,7 @@
 """
 Filename: calculate_mix_thread.py
 Author: Dipl.-Ing. (FH) Jonas Pfeiffer
-Date: 2024-07-31
+Date: 2024-09-10
 Description: Contains the threaded functionality function for calculating the heat generation mix.
 """
 
@@ -11,7 +11,7 @@ import traceback
 from PyQt5.QtCore import QThread, pyqtSignal
 
 from net_simulation_pandapipes.pp_net_time_series_simulation import import_results_csv
-from heat_generators.heat_generator_classes import Berechnung_Erzeugermix, optimize_mix
+from heat_generators.heat_generation_mix import Berechnung_Erzeugermix, optimize_mix
 
 class CalculateMixThread(QThread):
     """
